Card Display: Fix log spam for certain y positions
The log would be spammed with "Failed to parse component: []
java.lang.NullPointerException" whenever no spaces to align the y
coordinate where needed.

This was because the y spaces text component would be an empty jsonArray
in that case and Minecraft apparently doesn't like that.

This fix simply always adds an empty string to that array.