Timer: Timer now uses the worldborder of the lobby
Minecraft only cares about the worldborder in the overworld. All
dimensions use that worldborder.

Executing a worldborder command in any other dimension modifies a
worldborder in RAM for that dimension. It can be used by commands like
the normal worldborder.

Such a worldborder is only kept in RAM and will be reset to 60,000,000
upon rejoin. Because of this, some extra sanity checks need to be done,
to prevent rejoins from messing things up.

The worldborder of the overworld must not be modified by other packs,
else the RAM worldborder will be influenced by that automatically.