Spawn: Fix minor bug and qol improvements
There was a bug causing the spawn function to be executed thrice,
because move_x/0 and move_z/0 had the commands in the wrong order.

Additionally, some feedback during the spawn generation has been added
and a bigger area around spawn is forceloaded to hopefully make loading
in a bit quicker.