Card Generation: Move spawn location random call to here
This moves the spawn location random int call from game/start to
card_generation.

This makes the system more robust, even if the internal state of random
gets modified after the generation, for example due to a reload or
because the game is started multiple times on the same seed.