Util: Improvements to find_player_team
This changes find_player_team to no longer set the selected tag of the
current player's team. Instead, the current team is sorted to the end of
the list and can be selected by bingo:card teams[-1].

This removes the reponsibility to unselect the current team from the
caller.

This should also improve performance, since minecraft can do a direct
index access instead of having to search the correct entry in the array.

Additionally, the tag bingo.in_current_team is now handled by that
method as well and does not need to be removed by the caller at the end.