Item Detection: Copy selected item & simplify clear logic
With this, instead of marking the item on the card directly with
selected: true, this now copies the current item to a temporary storage.

Because of this, the clear logic for tagging players that need clearing
had to change, since we cannot store player IDs along with the item
anymore.

Instead there is now a seperate list of lists of player ids. Each
sub-list appears in the order the items where received and the commands
for cleaing will be executed in.

This also means, it's no longer neccessary to mark the item during the
clear process, and all items in the registry only need one clear command
stored now.

This incidentally also fixes #7