Game: All lcg calls before the first coordinate calculation
This adds all 6 lcg calls before the first coordinate calculation.

Yes this is terribly doublicated code everything is inlined and it's
hard to manage. But I want to keep it as much inlined as possible to
save as many commands as possible.

In the end, we can just treat this mess as a black box anyway. Unless
Minecraft code changes we don't need to touch it ever again.