Game: Fix emerald detection
This makes emerald detection reliable.

The previous approach by implementing biome fuzzing is removed again.
Turns out chunk coordinate 10 10 is extremely good for detecting the
correct biome. Just check 8 different y spots, and we can be extremely
sure that if most of them are mountains, it's an emerald chunk.

Fixes #11