Init: Register items with detect command and make system use those
This removes the advancements that were previously used for item
detection.

Instead, each item is now registered with a detect command that is
copied to a command block when a card is generated.

When adding a new item, the only thing that has to be done now in the
datapack is to register it.

Fixes #15
Closes #22 (not sure this is actually fixed, but since the system is
entirely different now, I'd want to make a new ticket if it's still
unreliable)
Resolves #31