Lobby: Move 2 up and make floor border 1 higher
This moves the entire lobby up by 2 and generates 3 layers of barriers
in the worldgen settings instead of 1.

Initially, the only purpose was to have barriers underneath of floor to
prevent accidental void holes, but in the process I found that a floor
border that is 2 deep looks kinda cool, so I decided to make that change
as well.