Increase alpha channel for glyph size forcers
Minecraft likes to automaitcally determine the width of characters. This
is a good feature when you use fonts for ... well text, but it's bad for
custom hud shenenigans.

To avoid that, there are some transparent pixels on both the left and
the right edge of each character, to ensure all item icons on the bingo
card share the same width.

However, for some reason OptiFine has a threshold for transparent pixels
and only will consider pixels above a certain translucency for
determining the glyph's width.

This increases the alpha channel so that even bad mods like OptiFine
work with card display.

Fixes #8