Preparations for new item management
This adds tags to categories and allows expansion packs to select
activeTags to toggle categories on and off.

This also manages items in one list with a global weight derived from
category-local weights.
This makes the whole model more sane mathematically and should make it
so items in multiple categories don't screw with probabilities.

Card generation still needs to be updated though, so this is currently
not used.