Rewrite card generation to use new system
Card generation used to pick categories and then a random item from that
category. This had some wonky implementation details, for example
categories that had already been picked were not properly removed in all
cases, leading to a system that has an uneven random distribution over
categories.

With the new system, all items are flattened into one array. When an
item is picked, all items that have at least one category in common with
that item are then removed.

The last commit added a way to dynamically change active categories and
calculated a global item weight based on all categories the item is in
to still ensure each category is picked with an equal chance and the
category-local weight is still applied.

If an item is in multiple categories, an average is taken.

With this there is also a cached value containing the total weight of
all currently active items, removing some overhead from card gen.

For expansion packs, you can now change the storage bingo:items
activeTags and then call the function bingo:util/apply_active_item_tags
to apply your changes.

The category definition now includes a tag array. Only categories that
share a tag with the activeTags array will be considered.

By default, the only active tag is the bingo:default tag.

Resolves #72
Fixes #74