Prevent gamemode switch and implement trigger
With this, manually switching the gamemode while in game is prevented,
since there are certain key combinations that could be used
accidentally.

Instead there is now a trigger for switching to spectator mode which
also allows using all other gamemodes if you are an operator.

This also makes spectator mode accessible for non-operators on a server.

Resolves #70