Add Bingo FX sound slider, custom card gen sound
This changes all custom Bingo sounds to use the "voice" slider and
renames this slider to "Bingo FX".

Furthermore, lobby item frames are made silent. Item rotate and removal
is still imitaed as it would be per usual, but the card gen sound is
customized now and plays a single sound rather than 25 sounds of item
frames getting filled