New pregen
Pregen was broken due to breaking changes to /forceload.

This implements a new version of pre-gen which works across multiple
ticks even when not all chunks are loaded yet.

This works by moving to the last chunk based on a scoreboard value
holding the offset to the actual spawn.

This also uses a new spiral-order for chunks to generate. This will
facilitate custom settings for the area to pre-generate in the future.

For now, the pregen-area is the same as Fetchr 5.0.

The lag calculation still needs more thought to possibly speed up
pregen. It's quite slow compared to 5.0. Not sure yet if this is my
fault or Mojang's.

Still needs logic to freeze entities and unload chunks, especially when
there is a setting for the amount of chunks to load.

Skybox is not spawning yet. In fact, after pre-gen is completed, nothing
happens. This will be done in a following commit.