Switch to custom space provider font
Looks like the legacy approach broke in 1.19.4. This was sth I wanted to
update anyway, and I guess now is the time.

There is an updated version of AmberWat's pack, but we can use a more
lightweight version with only spacings which we actually need.

This adds a custom space font and replaces the dependency to the
negativce space font.