Automatically generate missing default strings
In order to override the text of the sound slider every language needs
to be overridden. Since Mojang likes to add the weirdest of languages
all the time, this is hard to keep up-to-date.

With this, the existing languages are fetched for the target Minecraft
version and generated to make sure everything is working correctly.

Also removes some unneeded or untranslated language files. The crowdin
thing will re-create the ones that are actually configured in crowdin.