Make brightness only apply close to the player
Remove the dimension brightness and instead place a light block close to
the player when they are in a cave.

The player is considered in a cave when they are no more than 5 blocks
below the ocean_floor heightmap at any of these relative positions:
- ~ ~ ~
- ~3 ~ ~
- ~7 ~ ~
- ~-3 ~ ~
- ~-7 ~ ~
- ~ ~ ~3
- ~ ~ ~7
- ~ ~ ~-3
- ~ ~ ~-7
- ~-5 ~ ~-5
- ~-5 ~ ~5
- ~5 ~ ~-5
- ~5 ~ ~5

The lightblock is attempted to be placed somewhere around the player. A
location is invalid for the lightblock when the position is neither air
nor water, or when one of the 6 adjcent blocks is a gravity block (sand or
gravel), or when one of the 6 adjcent blocks is air when the current
position is water, or the other way around, or when one of the 6 adjcent
blocks is lava.

The following positions relative from the player's feet are considered
in this order; if none of these positions are valid, no light block is
created.
- ~ ~1 ~
- ~ ~ ~
- ~ ~-1 ~
- ~1 ~1 ~
- ~ ~1 ~1
- ~-1 ~1 ~
- ~ ~1 ~-1
- ~ ~2 ~
- ~1 ~ ~
- ~ ~ ~1
- ~-1 ~ ~
- ~ ~ ~-1
- ~ ~-2 ~
- ~3 ~1 ~
- ~ ~1 ~3
- ~-3 ~1 ~
- ~ ~1 ~-3
- ~3 ~ ~
- ~ ~ ~3
- ~-3 ~ ~
- ~ ~ ~-3