Rewrite pre-generation and spawning in behavior
This re-writes pre-generation to work entity-less and also successfully
works around /suggestload and freezes entities properly upon loading,
and unfreezes them when the game starts.

Furthermore, players get a preview of spawn much earlier now, which also
enables join logic and being on a server without joining a game, and
possibly joining an ongoing game as a spectator in the future.

Fixes #160
Resolves #154