Make sure ow is loaded for command block check
This fixes an issue when using Paper. Paper is still not supported and
seems to not keep forceloaded chunks when restarting the server.

Pre-gen also sometimes seems to freeze, also related to forceload not
working properly.

Other than than that, it actually seems fine by now for the most part.