Improve neunscript since/until
With this, when a file would be completely empty, it is not written at all.
This means in essence, when a since or until command does leave a completely
empty file for an earlier version, it is not written.

Furthermore, if, unless, since and until now only uncomment if the line
immediately after them is a comment. This allows for commented lines in a
section that is conditionally kept