Update lobby for 1.20.5, add join as spectator and points race
When a game is already ongoing, you can join as a spectator now instead of
being stuck in the lobby.

A new points race mode has been added. When active, a goal announcing the
team(s) with the most amount of points after a specified amount of time, is
active.