Fix item get message showing an outdated item count
The item get message was being shown before fetching and updating the item
count, meaning that the number related to the last thing that should have
been shown was shown instead.

In singleplayer, that meant the count was generally off by one, in multi player
the item count of the last team that triggered an update would be shown
instead.

Fixes #244