Fix neunscript not removing properly from previous versions
When there was a file that is supposed to be removed from earlier verions, the
file needs to be added with an overlay to make sure it only exists in the new
verion. This is a needed exception to the rule tha all current files should be
present in the main data dir.