neun_script: Create zip file in ram in favor of tmp directories, switch nbt lib
This has the nice side effect that the created zip files now have a consistent
hash, at least on the same device. It depends on file modification dates which
will be different for everyone. But for me, it doesn't matter how often I run
the build script, e.g. the resource pack sha1 will always be consistent.